feat: setup project
