+++ package.json
@@ -1,21 +1,33 @@
 {
-    "name": "ws_task",
+    "name": "battleshpip",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "nodemon --exec ts-node ./src/index.ts",
+        "start:dev": "nodemon ./src/index.ts",
+        "start:prod": "tsc && node ./dist/index.js",
+        "format": "prettier --write \"src/**/*.{ts,tsx,js,jsx,css,scss}\"",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
-    "author": "Andrei Auchynnikau",
+    "author": "Sergey Volkov",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
+        "@types/node": "^20.11.19",
         "nodemon": "^2.0.16"
     },
     "dependencies": {
+        "@types/ws": "^8.5.10",
+        "@typescript-eslint/eslint-plugin": "^7.0.2",
+        "@typescript-eslint/parser": "^7.0.2",
+        "dotenv": "^16.4.5",
+        "eslint": "^8.56.0",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-prettier": "^5.1.3",
+        "prettier": "^3.2.5",
+        "ts-node": "^10.9.2",
+        "typescript": "^5.3.3",
         "ws": "^8.8.0"
     }
 }
